+++ functions/card.js
@@ -1,18 +1,19 @@
 const main = require('../index');
 const Discord = require('discord.js');
+const _ = require('lodash');
 const fetchCard = require('./fetch').fetchCard;
 const buildAttachment = require('./buildAttachment').buildAttachment;
-const sets = {341: 'CotA', 435: 'AoA'};
+const sets = require('../data').sets;
 
-
-const card = async (msg, params, client, lang, set) => {
-	const data = fetchCard(params.join(' '), lang, set);
+const card = async (msg, params, flags) => {
+	const data = fetchCard(params.join(' '), flags);
 	const embed = new Discord.RichEmbed();
 	if (data) {
 		const title = `${data.card_number}.png`;
-		const attachment = await buildAttachment([data], title, lang);
+		const attachment = await buildAttachment([data], title, flags);
+		const set = _.get(sets.filter(set => data.expansion === set.set_number), '[0].flag', 'ERROR');
 		embed.setColor('031763')
-			.setDescription(`**[${data.card_title} #${data.card_number} ${sets[data.expansion]}](https://keyforge-compendium.com/cards/${data.card_number}?powered_by=archonMatrixDiscord)** `)
+			.setDescription(`**[${data.card_title} #${data.card_number} ${set}](https://keyforge-compendium.com/cards/${data.card_number}?powered_by=archonMatrixDiscord)** `)
 			.attachFile(attachment)
 			.setImage(`attachment://${title}`)
 			.setFooter(`Link provided by KeyForge Compendium`);