+++ handlers.js
@@ -1,8 +1,6 @@
 const knownCommands = require('./functions/index');
 const commandPrefix = require('./config').prefix;
 const version = require('./package').version;
-const langs = require('./data').langs;
-const sets = require('./data').sets;
 const sites = ['https://www.keyforgegame.com/deck-details/',
 	'https://decksofkeyforge.com/decks/',
 	'https://keyforge-compendium.com/decks/',
@@ -14,23 +12,16 @@ const sites = ['https://www.keyforgegame.com/deck-details/',
 const onMessage = (msg, client) => {
 	if (msg.author.bot) return; // Ignore messages from the bot
 
-	if (sites.some(a=>msg.content.startsWith(a))) {
+	if (sites.some(a => msg.content.startsWith(a))) {
 		const [site] = sites.filter(a => msg.content.startsWith(a));
-		msg.content = `${commandPrefix}d ${msg.content.replace(site, '').slice(0,36)}`
+		msg.content = `${commandPrefix}d ${msg.content.replace(site, '').slice(0, 36)}`
 	}
 	// This isn't a command since it has no prefix:
 	if (!msg.content.includes(commandPrefix) && !msg.content.match(/[\[\]{}]/g)) return;
-	let params = [], commandName, lang = 'en', set, message = msg.content.toLowerCase();
+	let params = [], commandName, message = msg.content.toLowerCase(), flags;
 
-	if (message.split(' ').some(a => a.startsWith('-'))) {
-		const flags = message.split(' ').filter(a => a.startsWith('-'));
-		flags.forEach(flag => {
-			flag = flag.slice(1);
-			if (langs.includes(flag)) lang = flag;
-			if (Object.keys(sets).includes(flag)) set = sets[flag];
-		});
-		message = message.split(' ').filter(a => !a.startsWith('-')).join(' ');
-	}
+	flags = message.split(' ').filter(a => a.startsWith('-')).map(flag => flag.slice(1));
+	message = message.split(' ').filter(a => !a.startsWith('-')).join(' ');
 
 	if (message.includes(commandPrefix)) {
 		if (message.startsWith(commandPrefix)) params = message.substr(1).split(' ');
@@ -75,8 +66,8 @@ const onMessage = (msg, client) => {
 
 	// If the command is known, let's execute it:
 	if (commandName in knownCommands) {
-		console.log(`${msg.author.username}, ${commandName}, ${params}, ${new Date()}`);
-		knownCommands[commandName](msg, params, client, lang, set);
+		console.log(`${msg.author.username}, ${commandName}, ${params}, ${flags}, ${new Date()}`);
+		knownCommands[commandName](msg, params, flags, client);
 	}
 
 };