+++ changelog.md
@@ -1,3 +1,6 @@
+# 1.2.7
+    added flags const
+    new loose card search
 # 1.2.6
     -adding in legacy card images for RH